Improve test coverage and reduce circular dependencies
